Refactor/package structure

+++ src/ui/src/components/AmberGlobalAdmin.vue
@@ -1,15 +1,23 @@
 <script setup lang="ts">
 import {ref, onMounted} from "vue"
-import { AmberClient, type Tenant, type UserDetails} from "amber-client"
+import { AmberClient, type Tenant} from "amber-client"
+import { globalTenant } from "../../../shared/src"
+import { uiHelper} from "@/state"
+import { state } from "@/state";
 var props = defineProps<{amberClient: AmberClient}>();
 
 var allTenants = ref<Tenant[]>([]);
 var newTenantName = ref("");
 var newTenantId = ref("");
-var show = ref(false);
 var adminApi = props.amberClient.getGlobalAdminApi()!;
 var deleteTenant = async (tenantId :string)=>{
  
+    if (!(await uiHelper.confirmDialog("Are you sure you want to delete tenant " + tenantId + "?")))
+    {
+      console.log("User cancelled delete tenant");
+      return;
+    }
+
     var result = await adminApi.deleteTenant(tenantId);
     if (result.success)
     {
@@ -31,24 +39,31 @@ var refreshTenants = async ()=>{
 onMounted(async ()=>{
   await refreshTenants();
 });
+
+var createLinkToTenantApp = (tenantId:string) :string =>
+{
+  var targetUrl = state.uiConfig.loginTargetUrl || "/";
+  targetUrl = targetUrl.replace('{tenant}', tenantId);
+    
+  return targetUrl;
+};
 </script>
 <template>
   <v-container>
   <v-row>
-      <h2>You are admin for all tenants
-        <v-btn v-if="!show" icon="mdi-menu-down" @click="show = true"></v-btn>
-        <v-btn v-if="show" icon="mdi-menu-up" @click="show = false"></v-btn>
+      <h2>Manage tenants
       </h2>
     </v-row>
-    <v-row v-if="show">
+    <v-row >
     <v-card width="100%">
-      <v-card-title>Manage tenants</v-card-title>
       <v-card-text>
         <v-list>
           <v-list-item v-for="tenant in allTenants" :key="tenant.id">
             <v-list-item-title>{{tenant.name}} [{{ tenant.id }}]</v-list-item-title>
             <v-list-item-action>
-              <v-btn @click="deleteTenant(tenant.id)">Delete</v-btn>
+              <v-btn icon="mdi-delete-outline" @click="deleteTenant(tenant.id)" title="delete" v-if="tenant.id != globalTenant"></v-btn>
+              <v-btn icon="mdi-application-edit" title="manage" :href='"admin?tenant=" + tenant.id'></v-btn>
+              <v-btn icon="mdi-open-in-new" title="visit" :href='createLinkToTenantApp(tenant.id)'></v-btn>
             </v-list-item-action>
           </v-list-item>
         </v-list>
